refactor(eloquent): extract ModelInstancePreparer from the builder
The builder held two domains: mirroring Model::__construct()'s
initializer lifecycle on a constructor-less instance, and computing
ModelMetadata sections from Psalm storage. The lifecycle half encodes
the fragile Laravel/PHP facts (bootTraits discovery branches, the 12.22
boundary, PHP-8.4-vs-8.5 getMethods() ordering, framework-vs-user source
discrimination), has no Psalm dependency, and is where the recent churn
landed. Extracted, it is unit-testable without warming a registry.
Mirrors the earlier CustomTypeDetector extraction.

Behaviour-preserving: every moved body is token-identical to its
original and no existing test expectation changed. The extracted
lifecycle also gains its own unit test, driving prepare() directly with
no registry or Codebase in the way.

The moved COMMENTS are condensed rather than relocated verbatim. Every
non-obvious fact survives — the PHP-version ordering, the source-file-
not-declaring-class trap and its trailing-slash guard, the 12.22 gate,
the boot-hang rationale, the reflection-not-dynamic-call trap, the known
gaps, the coverage shapes. Prose and restatement do not.

- replayInitializers() becomes the sole public entry point, prepare().
  It already ran the post-walk #[Connection]/#[Table] phase, so the one
  call site moves intact; the computeSection guard, failure recording
  and section gating stay with the builder, which is what lets a
  throwing initializer keep degrading exactly four sections.
- classAttribute() had no builder call sites outside the moved cluster,
  so it moves wholesale and stays private. declaresOwnProperty() and
  rawTableIsNull() move with their only caller, applyTableAttribute().
- filterStringList() stays in the builder (8 call sites in other
  sections); the preparer keeps a private copy rather than call across
  and point itself back at the class that depends on it.

Re-reading each comment while condensing turned up two false claims,
both corrected rather than carried over:

- classAttribute() said a throwing attribute constructor "reaches the
  runtime-configuration boundary". It never did; the failure has always
  landed in whichever section wraps the replay, and the class docblock
  already states that contract. Clause dropped.
- applyConcernMirror() said the mirrors were "verified 1:1" and that
  HasTimestamps no-ops because nothing it sets is stored. usesTimestamps
  IS stored via computeTraitFlags(), and no mirror applies
  #[WithoutTimestamps] / #[Table(timestamps:)], so such a model records
  the wrong value. Pre-existing; filed as #1276 and documented as a
  KNOWN GAP.
