+++ tests/Application/app/Models/AttributeSerializableModel.php
@@ -13,7 +13,7 @@ use Illuminate\Database\Eloquent\Model;
  * Serialization archetype driven by Laravel 13.0+ PHP class attributes (`#[Appends]` / `#[Hidden]`)
  * instead of `$properties`. The harness runs no migrations, so the shape comes entirely from the
  * appended, accessor-backed attributes — letting AttributeConfigShapeTest.phpt assert end-to-end that
- * `replayInitializers()` feeds the serialized shape (and that `#[Hidden]` drops an appended key).
+ * `ModelInstancePreparer::prepare()` feeds the serialized shape (and that `#[Hidden]` drops an appended key).
  *
  * Modern protected `Attribute` accessors keep the legacy public-accessor lint (`PublicModelAccessor`)
  * quiet. The `#[*]` classes only exist from Laravel 13.0, so the phpt is SKIPIF-gated below that line.
